chore: move keyring impl to configuration package
This prevent a circular import between the configuration and controllers packages.
